Reorganize modules and doc for Forms
This organization I think makes more sense and separates data-specific
wrapping modules, like 'MatrixForm', from generall purpose $PrintForm
output.

Our Mathics3-specfic forms have been isolated.

Text has been gone over.
