+++ mathics/builtin/forms/__init__.py
@@ -1,10 +1,9 @@
 r"""
 Forms of Input and Output
 
-A <i>Form</i> format specifies the way Mathics Expression input is read or output \
-written.
+A <i>Form</i> specifies the way Mathics3 Expressions are formatted and displayed.
 
-The variable <url>:\$OutputForms':
+The variable <url>:$OutputForms:
 /doc/reference-of-built-in-symbols/forms-of-input-and-output/form-variables/\$outputforms/</url> \
 has a list of Forms defined.
